backend/requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
requests
openai
google-genai
